+++ src/array_subset/array_subset_iterators/contiguous_linearised_indices_iterator.rs
@@ -5,17 +5,15 @@ use crate::{
     array_subset::{ArraySubset, IncompatibleArraySubsetAndShapeError},
 };
 
-use super::ContiguousIndicesIterator;
+use super::{contiguous_indices_iterator::ContiguousIndices, ContiguousIndicesIterator};
 
-/// Iterates over contiguous linearised element indices in an array subset.
-///
-/// The iterator item is a tuple: (linearised index, # contiguous elements).
-pub struct ContiguousLinearisedIndicesIterator<'a> {
-    inner: ContiguousIndicesIterator,
-    array_shape: &'a [u64],
+/// TODO
+pub struct ContiguousLinearisedIndices {
+    inner: ContiguousIndices,
+    array_shape: Vec<u64>,
 }
 
-impl<'a> ContiguousLinearisedIndicesIterator<'a> {
+impl ContiguousLinearisedIndices {
     /// Return a new contiguous linearised indices iterator.
     ///
     /// # Errors
@@ -23,9 +21,9 @@ impl<'a> ContiguousLinearisedIndicesIterator<'a> {
     /// Returns [`IncompatibleArraySubsetAndShapeError`] if `array_shape` does not encapsulate `subset`.
     pub fn new(
         subset: &ArraySubset,
-        array_shape: &'a [u64],
+        array_shape: Vec<u64>,
     ) -> Result<Self, IncompatibleArraySubsetAndShapeError> {
-        let inner = subset.iter_contiguous_indices(array_shape)?;
+        let inner = subset.contiguous_indices(&array_shape)?;
         Ok(Self { inner, array_shape })
     }
 
@@ -35,8 +33,8 @@ impl<'a> ContiguousLinearisedIndicesIterator<'a> {
     ///
     /// `array_shape` must encapsulate `subset`.
     #[must_use]
-    pub unsafe fn new_unchecked(subset: &ArraySubset, array_shape: &'a [u64]) -> Self {
-        let inner = subset.iter_contiguous_indices_unchecked(array_shape);
+    pub unsafe fn new_unchecked(subset: &ArraySubset, array_shape: Vec<u64>) -> Self {
+        let inner = subset.contiguous_indices_unchecked(&array_shape);
         Self { inner, array_shape }
     }
 
@@ -45,6 +43,32 @@ impl<'a> ContiguousLinearisedIndicesIterator<'a> {
     pub fn contiguous_elements(&self) -> u64 {
         self.inner.contiguous_elements()
     }
+
+    /// Create a new serial iterator.
+    #[must_use]
+    pub fn iter(&self) -> ContiguousLinearisedIndicesIterator<'_> {
+        <&Self as IntoIterator>::into_iter(self)
+    }
+}
+
+impl<'a> IntoIterator for &'a ContiguousLinearisedIndices {
+    type Item = (u64, u64);
+    type IntoIter = ContiguousLinearisedIndicesIterator<'a>;
+
+    fn into_iter(self) -> Self::IntoIter {
+        ContiguousLinearisedIndicesIterator {
+            inner: self.inner.into_iter(),
+            array_shape: &self.array_shape,
+        }
+    }
+}
+
+/// Iterates over contiguous linearised element indices in an array subset.
+///
+/// The iterator item is a tuple: (linearised index, # contiguous elements).
+pub struct ContiguousLinearisedIndicesIterator<'a> {
+    inner: ContiguousIndicesIterator<'a>,
+    array_shape: &'a [u64],
 }
 
 impl Iterator for ContiguousLinearisedIndicesIterator<'_> {
@@ -58,6 +82,14 @@ impl Iterator for ContiguousLinearisedIndicesIterator<'_> {
 
     fn size_hint(&self) -> (usize, Option<usize>) {
         self.inner.size_hint()
+    }
+}
+
+impl DoubleEndedIterator for ContiguousLinearisedIndicesIterator<'_> {
+    fn next_back(&mut self) -> Option<Self::Item> {
+        self.inner
+            .next_back()
+            .map(|(indices, elements)| (ravel_indices(&indices, self.array_shape), elements))
     }
 }